backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg==3.1.13
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
python-magic==0.4.27
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2